edit user and products migration

+++ database/migrations/2023_03_17_045242_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->timestamps();
             $table->string('productName');
-            $table->bigInteger('price');
+            $table->double('price');
         });
     }
